requirements.txt
accelerate==0.33.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
biopython==1.83
biotite==0.40.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
fair-esm==2.0.0
filelock==3.13.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.4.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
lightning-utilities==0.11.6
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.0
numba==0.58.1
numpy==1.24.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.0.3
pillow==10.2.0
psutil==6.0.0
pyg-lib==0.4.0+pt22cu121
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
sympy==1.12
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.1+cu121
torch_cluster==1.6.3+pt22cu121
torch_geometric==2.5.3
torch_scatter==2.1.2+pt22cu121
torch_sparse==0.6.18+pt22cu121
torch_spline_conv==1.2.2+pt22cu121
torchaudio==2.2.1+cu121
torchmetrics==1.2.1
torchvision==0.17.1+cu121
tqdm==4.66.5
transformers==4.44.0
triton==2.2.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
zipp==3.20.0
